Use MyST-NB to render jupyter notebook and markdown files
MyST-NB parses jupyter notebooks and markdown in Sphinx, thereby replacing nbsphinx (for .ipynb files) and recommonmark (for .md files). The powerful MyST (Markedly Structured Text) parser also allows much of the documentation to be written in Markdown instead of reStructuredText, which can make it easier for new folks wanting to make contributions to the documentation.