Adjust tests to account for renaming legacy test module
The module was renamed by a Ruff rule, but there are tests that depend
on the name of the module. Those tests are updated to the new module
name.

The legacy module was named without the `test_` prefix so that it
wouldn't be collected by the test runner. Because it does contain tests,
the prefix is added back, and now we tell the test runner to skip the
module and why.