Create new base class for query objects
`BaseQuery` contains the functionality we expect to use with
harmony/earthaccess. The `Query` class now just has the CMR/EGI-specific codes
and elements that will not be supported in v2 (or still need to be figured out,
like `cycles` and `tracks`

TODO: do we need `BaseQuery` _and_ `GenQuery`? `GenQuery` doesn't appear to be
used elsewhere.